Metadata handling for all PDF endpoints (#1894)
* Add image support to multi-tool page

Related to #278

* changes to support image types

* final touches

* final touches

* final touches


* final touches


* final touches


* final touches


* final touches


* final touches


* final touches


* Update translation files (#1888)


* final touches


---------

+++ src/main/java/stirling/software/SPDF/controller/api/converters/ConvertWebsiteToPDF.java
@@ -6,6 +6,10 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.pdfbox.pdmodel.PDDocument;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,6 +20,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 import stirling.software.SPDF.model.api.converters.UrlToPdfRequest;
+import stirling.software.SPDF.service.CustomPDDocumentFactory;
 import stirling.software.SPDF.utils.GeneralUtils;
 import stirling.software.SPDF.utils.ProcessExecutor;
 import stirling.software.SPDF.utils.ProcessExecutor.ProcessExecutorResult;
@@ -25,6 +30,15 @@ import stirling.software.SPDF.utils.WebResponseUtils;
 @Tag(name = "Convert", description = "Convert APIs")
 @RequestMapping("/api/v1/convert")
 public class ConvertWebsiteToPDF {
+
+    private static final Logger logger = LoggerFactory.getLogger(ConvertWebsiteToPDF.class);
+
+    private final CustomPDDocumentFactory pdfDocumentFactory;
+
+    @Autowired
+    public ConvertWebsiteToPDF(CustomPDDocumentFactory pdfDocumentFactory) {
+        this.pdfDocumentFactory = pdfDocumentFactory;
+    }
 
     @PostMapping(consumes = "multipart/form-data", value = "/url/pdf")
     @Operation(
@@ -46,12 +60,12 @@ public class ConvertWebsiteToPDF {
         }
 
         Path tempOutputFile = null;
-        byte[] pdfBytes;
+        PDDocument doc = null;
         try {
             // Prepare the output file path
             tempOutputFile = Files.createTempFile("output_", ".pdf");
 
-            // Prepare the OCRmyPDF command
+            // Prepare the WeasyPrint command
             List<String> command = new ArrayList<>();
             command.add("weasyprint");
             command.add(URL);
@@ -61,16 +75,23 @@ public class ConvertWebsiteToPDF {
                     ProcessExecutor.getInstance(ProcessExecutor.Processes.WEASYPRINT)
                             .runCommandWithOutputHandling(command);
 
-            // Read the optimized PDF file
-            pdfBytes = Files.readAllBytes(tempOutputFile);
-        } finally {
-            // Clean up the temporary files
-            Files.deleteIfExists(tempOutputFile);
-        }
-        // Convert URL to a safe filename
-        String outputFilename = convertURLToFileName(URL);
+            // Load the PDF using pdfDocumentFactory
+            doc = pdfDocumentFactory.load(tempOutputFile.toFile());
 
-        return WebResponseUtils.bytesToWebResponse(pdfBytes, outputFilename);
+            // Convert URL to a safe filename
+            String outputFilename = convertURLToFileName(URL);
+
+            return WebResponseUtils.pdfDocToWebResponse(doc, outputFilename);
+        } finally {
+
+            if (tempOutputFile != null) {
+                try {
+                    Files.deleteIfExists(tempOutputFile);
+                } catch (IOException e) {
+                    logger.error("Error deleting temporary output file", e);
+                }
+            }
+        }
     }
 
     private String convertURLToFileName(String url) {
